pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "patchhawk"
version = "1.0.0"
description = "RL-powered supply-chain vulnerability detection & auto-patching"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    # Core
    "openenv-core>=0.2.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    # Validation / Docker
    "docker>=7.0.0",
    # A2A server
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    # Testing
    "pytest>=8.0.0",
]

[project.optional-dependencies]
training = [
    "unsloth>=2024.0",
    "trl>=0.7.0",
    "transformers>=4.38.0",
    "torch>=2.1.0",
    "peft>=0.8.0",
    "datasets>=2.16.0",
    "wandb>=0.16.0",
]
dashboard = [
    "streamlit>=1.30.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["patchhawk*", "server*"]
